Move bits around
